+++ index.js
@@ -3,15 +3,12 @@
 // Dependencies
 // ===
 
-const BodyParser = require('body-parser')
+const bodyParser = require("body-parser")
 
 const CookieParser = require("cookie-parser")
 
 // web server
 const Express = require("express")
-
-// better than debugger
-const Pry = require("pryjs")
 
 // HTTP requests
 const Request = require("request")
@@ -19,30 +16,21 @@ const Request = require("request")
 // Logging
 const Winston = require("winston")
 
+const cookieUtil = require("./lib/cookieUtil")
+const mixpanelUtil = require("./lib/mixpanelUtil")
+
 
 // Config
 // ===
 
 const MIXPANEL_API_HOST = process.env.MIXPANEL_API_HOST || "https://api.mixpanel.com"
 // e.g. token1234a,token5678b
-const MIXPANEL_TOKEN_WHITELIST = process.env.MIXPANEL_TOKEN_WHITELIST.split(",")
 const NODE_ENV = process.env.NODE_ENV || "development"
 const LOG_LEVEL = process.env.LOG_LEVEL || (NODE_ENV === "production" ? "info" : "debug")
-const COOKIE_PERSISTED_PARAMS = process.env.COOKIE_PERSISTED_PARAMS
-  ? process.env.COOKIE_PERSISTED_PARAMS.split(",")
-  : ["campaign", "creative", "placement", "referer", "utm_source", "utm_medium", "utm_campaign", "utm_content", "utm_term"]
-const PERSISTED_COOKIE_NAME = "metricProxy"
-const COOKIE_DOMAIN = process.env.COOKIE_DOMAIN
-const COOKIE_SECURE_ATTRIBUTE = (NODE_ENV === "production")
-const COOKIE_SIGNING_SECRET = NODE_ENV === "production"
-  ? process.env.COOKIE_SIGNING_SECRET
-  : null
-const COOKIE_TTL = 14 * 24 * 60 * 60 * 1000
 const PORT = process.env.PORT || 4000
 const USER_AGENT = `metric-proxy/${process.env.npm_package_version} (brave.com)`
 
-
-// Lib
+// Logging
 // ===
 
 var logger = new (Winston.Logger)({
@@ -52,153 +40,22 @@ var logger = new (Winston.Logger)({
   ]
 })
 
-// Mixpanel query string includes:
-// - Original request query string -- certain params
-// - Certain other query params, moved from top level into query data.properties
-// - Persisted params from previous requests, via cookies
-function buildMixpanelTrackQueryString(request, response) {
-  if (!request.query.data) {
-    throw "Query missing"
-  }
-
-  const dataString = decodeURI(Buffer.from(request.query.data, "base64").toString("utf-8"))
-  // HACK: Handle mixpanel iOS Swift 2.x library which sends single quoted JSON.
-  const data = JSON.parse(dataString.replace(/'/g, "\""))
-  if (!isValidMixpanelTrackData(data)) {
-    throw "Invalid data"
-  }
-
-  let queryString = { data }
-  // /track supports other params but not sure if we want them
-  const queryStringParams = ["img", "verbose"]
-  queryStringParams.forEach((value) => {
-    if (!request.query[value]) {
-      return
-    }
-    queryString[value] = request.query[value]
-  })
-
-  // Copy special params from query string
-  COOKIE_PERSISTED_PARAMS.forEach((key) => {
-    if (!request.query[key]) {
-      return
-    }
-    queryString.data.properties[key] = request.query[key]
-  })
-
-  // Restore persisted params
-  const restoredParams = restorePersistedParams(request, response)
-  if (Object.keys(restoredParams).length > 0) {
-    for (let key in restoredParams) {
-      if (queryString.data.properties[key]) {
-        continue
-      }
-      queryString.data.properties[key] = restoredParams[key]
-    }
-  }
-
-  logger.debug("API > Query/data:", queryString.data)
-  const returnDataString = JSON.stringify(queryString.data)
-  queryString.data = Buffer.from(returnDataString, "utf-8").toString("base64")
-  return queryString
-}
-
-function isValidMixpanelToken(token) {
-  if (!token) {
-    return false
-  }
-  // HACK: Handle mixpanel iOS Swift 2.x library which sends token as "Optional({token})"
-  return MIXPANEL_TOKEN_WHITELIST.some((whitelistToken) => {
-    return token.includes(whitelistToken)
-  })
-}
-
-// Log additional things in development environments.
-function debugLogger(request, response, next) {
-  logger.debug("-> Headers:", request.headers)
-  if (request.cookies) {
-    for (let cookieName in request.cookies) {
-      logger.debug(`-> Cookie: ${cookieName}:`, request.cookies[cookieName])
-    }
-  }
-
-  if (request.query.data) {
-    logger.debug("-> Query:", request.query)
-    const dataString = decodeURI(Buffer.from(request.query.data, "base64").toString("utf-8"))
-    // HACK: Handle mixpanel iOS Swift 2.x library which sends single quoted JSON.
-    const data = JSON.parse(dataString.replace(/'/g, "\""))
-    logger.debug("-> Query data:", data)
-  } else {
-    logger.debug("-> Query: (empty)")
-  }
-
-  if (request.body) {
-    logger.debug("-> Body:", request.body)
-    // HACK: Handle mixpanel iOS Swift 2.x library which sends single quoted JSON.
-    const dataBody = JSON.parse(request.body.replace(/'/g, "\""))
-    logger.debug("-> Body data:", dataBody)
-  } else {
-    logger.debug("-> Body: (empty)")
-  }
-
-  next()
-}
-
 // Log Request responses from Mixpanel
 function logRequestResponse(response) {
   logger.debug("<<", response.statusCode)
   logger.debug("<<", response.headers)
 }
 
-// Takes an example URL with tracking parameters and persists them to cookies.
-// Alternative to the mixpanel library, which persists things like utm_* as
-// Super Properties.
-// Useful where JS is unavailable like tracking pixels or click redirects.
-// Returns persisted params
-function persistParamsAsCookies(request, response) {
-  let cookieParams = {}
-  COOKIE_PERSISTED_PARAMS.forEach((key) => {
-    if (!request.query[key]) {
-      return
-    }
-    cookieParams[key] = request.query[key]
-  })
-  if (Object.keys(cookieParams).length === 0) {
-    return {}
-  }
-  let options = {
-    httpOnly: true,
-    expires: new Date(Date.now() + COOKIE_TTL),
-    secure: COOKIE_SECURE_ATTRIBUTE,
-    signed: !!COOKIE_SIGNING_SECRET
-  }
-  if (COOKIE_DOMAIN) {
-    options["domain"] = COOKIE_DOMAIN
-  }
-  logger.debug(`<< Cookie: ${PERSISTED_COOKIE_NAME}:`, cookieParams)
-  response.cookie(PERSISTED_COOKIE_NAME, cookieParams, options)
-  return cookieParams
-}
-
-function restorePersistedParams(request, response) {
-  if (!request.cookies[PERSISTED_COOKIE_NAME]) {
-    return {}
-  }
-  // TODO: More logics
-  return request.cookies[PERSISTED_COOKIE_NAME]
-}
-
-
 // App fn
 // ===
 
 // https://mixpanel.com/help/reference/http
-function mixpanelTrack(request, response) {
+function mixpanelTrackGet(request, response) {
   try {
     logger.info(`-> ${request.method} /track`)
     // Save certain query params across requests with cookies
-    persistParamsAsCookies(request, response)
-    const queryString = buildMixpanelTrackQueryString(request, response)
+    cookieUtil.persistParamsAsCookies(request, response)
+    const queryString = mixpanelUtil.buildMixpanelTrackQueryString(request, response)
     const mixpanelRequestOptions = {
       headers: {
         "User-Agent": USER_AGENT
@@ -229,29 +86,25 @@ function mixpanelTrack(request, response) {
   }
 }
 
-function isValidMixpanelTrackData(data) {
-  return (data.event && data.properties && isValidMixpanelToken(data.properties.token))
-}
-
 
 // App
 // ===
-
 
 // Express setup
 // ---
 var express = Express()
 express.disable("x-powered-by")
-if (COOKIE_SIGNING_SECRET) {
-  express.use(CookieParser(COOKIE_SIGNING_SECRET))
+if (cookieUtil.COOKIE_SIGNING_SECRET) {
+  express.use(CookieParser(cookieUtil.COOKIE_SIGNING_SECRET))
 } else {
   express.use(CookieParser())
 }
-express.use(BodyParser.text({type: "*/*"}))
+const urlencodedParser = bodyParser.urlencoded({ extended: false })
+express.use(urlencodedParser)
 if (LOG_LEVEL === "debug") {
-  express.use(debugLogger)
+  const {debugRequestLogger} = require("./lib/logUtil")
+  express.use(debugRequestLogger(logger))
 }
-
 
 // Routes
 // ---
@@ -259,8 +112,8 @@ express.get("/", (request, response) => {
   response.send("hello friend")
 })
 
-express.get("/track", mixpanelTrack)
-express.post("/track", mixpanelTrack)
+express.get("/track", mixpanelTrackGet)
+express.post("/track", mixpanelTrackGet)
 
 express.listen(PORT)
